Last attempt didn't work, but I was right about the bytes being wrong
I tried to do all of this myself, but I kept running into issues, so I just looked at PhacoxsInjector to see what they were doing. Based on what they have, my code looked like it was going down the wrong path so I based a lot of this code off of theirs and changed up some lines to work with ours and change some of the names.

I got this to work once, but my UWUVCI is borked af so I can't check against different roms or bases.